feat: response, request dto 경로 정리
