+++ src/main/java/com/springboot/petProject/dto/request/comment/CommentRequest.java
@@ -1,4 +1,4 @@
-package com.springboot.petProject.dto.request;
+package com.springboot.petProject.dto.request.comment;
 
 import jakarta.validation.constraints.Size;
 import lombok.Data;
