+++ src/main/java/com/springboot/petProject/dto/request/post/PostUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.springboot.petProject.dto.request;
+package com.springboot.petProject.dto.request.post;
 
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
@@ -6,7 +6,7 @@ import lombok.Data;
 
 @Data
 @AllArgsConstructor
-public class PostCreateRequest {
+public class PostUpdateRequest {
 
     @Size(min = 1, message = "Post should have atleast 1 characters")
     private String title;
