+++ src/main/java/com/springboot/petProject/dto/response/post/PostResponse.java
@@ -1,4 +1,4 @@
-package com.springboot.petProject.dto.response;
+package com.springboot.petProject.dto.response.post;
 
 import com.springboot.petProject.dto.PostDto;
 import com.springboot.petProject.util.DateUtil;
@@ -7,22 +7,20 @@ import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public class PostsResponse {
+public class PostResponse {
 
     private Integer postId;
     private String title;
     private String contents;
     private String nickname;
-    private int commentsCount;
     private String formattedRegisteredAt;
 
-    public static PostsResponse fromDto(PostDto post) {
-        return new PostsResponse(
+    public static PostResponse fromDto(PostDto post) {
+        return new PostResponse(
                 post.getPostId(),
                 post.getTitle(),
                 post.getContents(),
                 post.getNickname(),
-                post.getCommentsCount(),
                 DateUtil.formatTimestamp(post.getRegisteredAt())
         );
     }
